+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 CHANGELOG
 ---------
 
+## NEXT RELEASE
+ - fix #2987982 - Classes not saving on captioned images
+
 ## 8.x-1.0-beta4 (2018-02-21)
  - fix #2919949 - The originalUpcast trigger a false positif
  - fix #2930454 - Doesn't work with imce image button